feat: transaction status component
