+++ packages/extension-ui/src/components/Video.tsx
@@ -37,7 +37,7 @@ export default React.memo(
   styled(Video)<VideoProps>(
     () => `
     padding: 0;
-    width:100%;
+    width: 100%;
   `
   )
 );
